# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "CSE 341 Project 1",
        "description": "REST API for CSE 341 Project 1",
        "version": "1.0.0"
    },
    "host": "cse341-project-1-hvl8.onrender.com",
    "basePath": "/",
    "schemes": ["https"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
        "/contacts/": {
            "get": {
                "description": "",
                "produces": ["application/json"],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "firstName": {
                                    "example": "any"
                                },
                                "lastName": {
                                    "example": "any"
                                },
                                "email": {
                                    "example": "any"
                                },
                                "favoriteColor": {
                                    "example": "any"
                                },
                                "birthday": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/contacts/{id}": {
            "get": {
                "description": "",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "firstName": {
                                    "example": "any"
                                },
                                "lastName": {
                                    "example": "any"
                                },
                                "email": {
                                    "example": "any"
                                },
                                "favoriteColor": {
                                    "example": "any"
                                },
                                "birthday": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/": {
            "get": {
                "description": "",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "Contact": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "6830c1332164274e3bf8b9b3"
                },
                "firstName": {
                    "type": "string",
                    "example": "John"
                },
                "lastName": {
                    "type": "string",
                    "example": "Doe"
                },
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "favoriteColor": {
                    "type": "string",
                    "example": "blue"
                },
                "birthday": {
                    "type": "string",
                    "example": "1990-01-15"
                }
            }
        },
        "ContactInput": {
            "type": "object",
            "properties": {
                "firstName": {
                    "type": "string",
                    "example": "John"
                },
                "lastName": {
                    "type": "string",
                    "example": "Doe"
                },
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "favoriteColor": {
                    "type": "string",
                    "example": "blue"
                },
                "birthday": {
                    "type": "string",
                    "example": "1990-01-15"
                }
            }
        }
    }
}